Refactor user client
