+++ tests/fixtures/onboarding.py
@@ -1,0 +1,2 @@
+def get_eth_private_key():
+    return "50c8e358cc974aaaa6e460641e53f78bdc550fd372984aa78ef8fd27c751e6f4"
