+++ x10/signing/onboarding.py
@@ -1,14 +1,16 @@
 from dataclasses import dataclass
 from datetime import datetime, timezone
 from functools import cached_property
+from typing import NamedTuple
 
-from eth_account.messages import SignableMessage, encode_typed_data
+from eth_account.messages import SignableMessage, encode_defunct, encode_typed_data
 from eth_typing import ChecksumAddress
 from fast_stark_crypto import generate_keypair_from_eth_signature, pedersen_hash
 from fast_stark_crypto import sign as stark_sign
 
 from x10.core.types import SignMessageCallback
 from x10.models.account import AccountModel
+from x10.utils.date import utc_now
 
 register_action = "REGISTER"
 sub_account_action = "CREATE_SUB_ACCOUNT"
@@ -248,3 +250,18 @@ def get_sub_account_creation_payload(
         account_registration=registration_payload,
         description=description,
     )
+
+
+class RequestSignature(NamedTuple):
+    value: str
+    time: str
+
+
+def sign_api_request(request_path: str, sign_message: SignMessageCallback) -> RequestSignature:
+    now = utc_now()
+    now_as_string = now.strftime("%Y-%m-%dT%H:%M:%SZ")
+    l1_message = f"{request_path}@{now_as_string}".encode(encoding="utf-8")
+    encoded_l1_message = encode_defunct(l1_message)
+    l1_signature = sign_message(encoded_l1_message)
+
+    return RequestSignature(l1_signature, now_as_string)
